+++ src/engine/search/mod.rs
@@ -128,7 +128,7 @@ pub fn search(
 ) -> (Move, WhiteEval) {
     let mut state = SearchState::new();
 
-    let mut time_strategy = TimeStrategy::new(game, time_control);
+    let mut time_strategy = TimeStrategy::new(game, time_control, options);
     time_strategy.init();
 
     tt.new_generation();