+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Split eval tapering into its own module
 * Bundle midgame and endgame evals into a `PhasedEval` struct
 * Fix taking up more memory than needed when the transposition table is resized repeatedly
+* Remove the default 50ms move overhead and add a UCI option to configure it
 
 ## [2.1]